Initiall work to use class loading
