+++ gradle-plugins/cordform-common/src/main/java/net/corda/cordform/NetworkParametersGenerator.kt
@@ -1,0 +1,7 @@
+package net.corda.cordform
+
+import java.nio.file.Path
+
+interface NetworkParametersGenerator {
+    fun run(baseDirectory: Path, notaryMap: Map<String, Boolean>)
+}
